x86/defconfig: Add defconfigs
Add a defconfig for building based on Ubuntu, forward ported to 5.2, then
stripped down to eliminate unneeded things to speed up the build.  To be
used when testing against qemu.

Also add a defconfig for building based on Centos for testing on emulation.

These will later be amended to include MHI and other relevant components.

Change-Id: Ia80faa7e7da9f9908677dee953267e4578496588
Signed-off-by: Jeffrey Hugo <jhugo@codeaurora.org>