mhi_bus: core: Allow the primary event ring to process data
According to the MHI spec, the primary event ring is to be used to handle
MHI commands, such as open/close of channels.  However, that is in addition
to the responsibilities of a normal event ring, which would be processing
data.

Add support for the primary event ring to handle data events, instead of
dropping them.

Change-Id: I264f4483cc4a893834db76d9f34dd6544ca310eb
Signed-off-by: Jeffrey Hugo <jhugo@codeaurora.org>