kbuild: binrpm-pkg: Propagate O= to rpmbuild
If the user specifies O= to indicate a specific output directory for the
build, rpmbuild does not honor this, and will use its default, which could
be the user's home directory.  In cases where the user has limited home
directory space, this could cause the build to outright fail.

In the case of the binrpm-pkg target, redefine the top directory for output
to be what the user specified in O=, thus the user will find a "rpmbuild"
subdirectory in that location with all of the RPM artifacts.

This does not apply to rpm-pkg, since we already cannot handle creating
the source tarball out of tree.

Change-Id: Idfeb92f55289331dfce09c643a589c85a3197480
Signed-off-by: Jeffrey Hugo <jhugo@codeaurora.org>