mhi_bus: dev: uci: Use channel name instead of pipe
The device file created by uci looks like "/dev/mhi_a100_00.01.00_pipe_2"
This contains a lot of good information, such as the pcie device id and
address.  However, the "pipe_X" part of it is rather obtuse.  It requires
the userspace entity to have detailed internal knowledge of the device
it is interacting with, to know which pipe maps to which channel.

It would be more useful to use the channel name as that is meaningful to
the userspace client, allows the channel mapping to change, and removes
specialized knowledge from the client.

Change-Id: Iab76f3f554c22f76559eb0a241571e9cc0ef38c5
Signed-off-by: Jeffrey Hugo <jhugo@codeaurora.org>