qaic: Implement manage ioctl
The manage ioctl allows userspace to use the control path to a qaic device
for the purpose of sending commands.  There is a variety of commands, but
the primary purpose of sending commands is to initialize the device so
that the data path is usable as a hardware accelerator.  The commands are
modeled as a request/response mechanism to the firmware running on the
device.

Change-Id: Ib09628231365b7b4e96d239db2a9d3f3c0dc34db
Signed-off-by: Jeffrey Hugo <jhugo@codeaurora.org>