mhi_bus: core: Cleanup channels in reset after work flush
The work items could be creating a channel while reset is being processed.
The channel won't be cleaned up before the work is flushed, and will leave
a device registered.  This will cause an issue due to a sysfs file
conflict when the device comes back after the reset, and the channel is
re-discovered.  Fix this by ensuring all channels are cleaned up after
work items have been verified to be complete.

Change-Id: Ieb7ffc82a3438c0cab5d7c8dc0314bc3be9ca1a5
Signed-off-by: Jeffrey Hugo <jhugo@codeaurora.org>