qaic: Implement wait for execute ioctl
The wait for execute ioctl allows the user to wait until the execute
processing of a handle is complete.  When the ioctl returns, it signals
that the device is done with the memory associated with the handle, and the
handle can be reused for another execute operation to the device.

Change-Id: I59eafe9b7163c351f9a3c180b4c5b865c6a3ade2
Signed-off-by: Jeffrey Hugo <jhugo@codeaurora.org>