include/linux/mhi.h: Move struct mhi_device_id
Move struct mhi_device_id to mhi.h so that we can build mhi as an external
module against another tree.  If we leave it in mod_devicetable.h, we have
a build error we can't recover from since the verion in the other tree will
be used which does not have the struct mhi_device_id definition.

Change-Id: I2071c4035a2168d38977b7486437e80b492a764f
Signed-off-by: Jeffrey Hugo <jhugo@codeaurora.org>