qaic: Add deactivate support
Deactivate is the inverse of activate - a DMA Bridge channel is being
unassigned and the respective resources freed.  When the user requests a
deactivate, we block further access to the channel.  When the device
confirms the channel has been released, we release our internal resources.
Then the channel is ready for reuse.

Change-Id: Ibf6dd225b1e210ba241bf81d1c18c7cccf763600
Signed-off-by: Jeffrey Hugo <jhugo@codeaurora.org>