qaic: Fix race between execute and free mem handle
The user should quiesce activity on a mem handle before attempting to free
it.  If they don't, its possible that an execute request and a free mem
handle request for the same handle could race.  In this case, ensure that
nothing fatal happens - chances are the execute will go through, but the
backing memory won't be freed until after that occurs.

Change-Id: I5746675db7aa1129648edcb187c70caacb192f45
Signed-off-by: Jeffrey Hugo <jhugo@codeaurora.org>