qaic: Fix dbc struct corruption
We only want to save resources which have been allocated in processing a
manage ioctl.  Currently, we fail to check to see if any resources have
been allocated, so we do the save operation every time, which will corrupt
the dbc structures with garbage data.  Fix this by adding the required
check.

Change-Id: I4174b7cdb9886abc5420fb9673948d435eed5469
Signed-off-by: Jeffrey Hugo <jhugo@codeaurora.org>