qaic: Fix double list_del issue
If we sent a control message to the device on behalf of the userspace
process and are waiting for the response, the userspace process could be
interrupted (ie killed).  This results in a race between msg_xfer() and
qaic_mhi_dl_xfer_cb() which could result in list_del() being called on
the elem list twice.  This causes a list corruption splat.

The issue results from msg_xfer unconditionally calling list_del().  We
can fix this by detecting if the elem is still on the list, and only
calling list_del() in that case since both acceses to the elem are
protected by a mutex.

Change-Id: Idef57a16fa60ba9df56d956239cdeff29789cd4e
Signed-off-by: Jeffrey Hugo <jhugo@codeaurora.org>