mhi_bus: core: Add soc_reset sysfs to controller device
Most MHI implementations have a "mhi_soc_reset" register.  Writing to this
register triggers a SoC reset to the SoC behind the MHI implementation.
This is useful to recover non-responsive devices.  In many cases, a PCIE
reset such as SBR only affect the link, and not the device behind the link.

Allow a controller to indicate if it has this capability.  If so, provide
a sysfs file to trigger the reset.

Writing to the sysfs file triggers the reset, ie:
echo 1 > /sys/bus/mhi/devices/<controller device>/soc_reset

Change-Id: I961ebb485ac32804bf2750ac05b96eccb50adb46
Signed-off-by: Jeffrey Hugo <jhugo@codeaurora.org>