qaic: Add soc_reset sysfs
The aic100 device implementes mhi_soc_reset, which can be hooked into the
MHI bus via the controller so that a sysfs file is present which can
trigger this reset when the file is written to.  This provides a useful
mechanism for the system admin to reset the device, such as when there is
an error the driver has not detected.

Change-Id: Ic0f5cb88cf73ff5e202e67a0f47c53d22e0cb500
Signed-off-by: Jeffrey Hugo <jhugo@codeaurora.org>