qaic: Cleanup after a device is removed
If suprise link down occurs (the device takes down the link), the pci
remove function is called.  Since suprise link down can occur at any time,
we need to be methodical about cleaning up anything that might be in
progress.  We don't want to leak resources or cause a crash.

At the same time, we need to notify existing users that the device is gone
and prevent new ones from attempting to access the now non-existent device.
The best we can do is remove the char dev to prevent new users, however
existing ones will still have open file descriptors (FDs).  Its not
possible to force the users to close the open FDs, or to yank the FDs from
the users, so we have to make it so that any stale FDs just return error
for any operation and hope that eventually users get the message that they
need to close the FD.

Change-Id: I184845573476b3c66e76df530d0f11168ea2430f
Signed-off-by: Jeffrey Hugo <jhugo@codeaurora.org>