qaic: Add module parameter to tune default of wait exec ioctl timeout
The wait_for_exec ioctl has a default timeout of 5 seconds.  Some platform
stress testing is expected to run for multiple hours, which far exceeds
the timeout of 5 seconds.

Provide a module parameter to allow the timeout value to be tuned as
needed for such testing.  The value is specified in milliseconds, and has
an effective upper bound of 2^31-1, or roughly 24 days.  Values greater
than this are intrepreted as "infinite".

Change-Id: Ief8166778bc3ea4654860561be769be56e38fd51
Signed-off-by: Jeffrey Hugo <jhugo@codeaurora.org>