mhi_bus: core: Unlock channel mutex during driver remove
The channel mutex is held while the client driver is notified that the
device is being removed.  This can cause a deadlock as the client driver
may need to quiesce activity during the device remove process, and that
activity is attempting to use the channel, so it would be waiting on the
mutex.  Since the channel is already marked as disabled by the time the
client driver is notified via device remove, any new or pending activity
will exit early.  Therefore we should unlock the channel mutex during
the client's notification to avoid possible deadlock scenarios.

Change-Id: I84affef7726f4ff7346d70e8a25255dff7cfc40a
Signed-off-by: Jeffrey Hugo <jhugo@codeaurora.org>