qaic: Fix chardev corrupt list_del issue
Open inodes are put on a list in the cdev struct for a chardev.  When the
inode is freed it is removed from the list.  When all references to the
cdev are released, the list is cleaned out.  If we destroy a chardev, but
there is an inode left active, the list is still in use and the cdev
memory is still required.  However if the same device comes back, we reuse
the cdev memory and re-init it which corrupts the list which is still in
use.  Finally, when the inode is released, we get a corrupt list_del
splat.  Fix this by using dynamic cdev instances instead of a single
static one, thus the "destroyed" cdev is free to hang around until
unneeded, and creating its replacement won't cause corruption.  The cdev
framework will make sure the memory gets freed when no longer needed.

Change-Id: Icd2ad3e54ccbcbc3b59749577f4e4f3154cbade2
Signed-off-by: Jeffrey Hugo <jhugo@codeaurora.org>