qaic: Wait for terminate response
The assumption that we can skip the terminate response has broken.  We
need to give the device time to clean up its use of resources that we may
have assigned to it, before we go and clean up those reources ourselves.
Therefore before cleaning up an exiting user, we need to wait for the
terminate response so that we've given the device the oppertunity to clean
up as well.

This has the side benefit of addressing a race condition from reusing
reciently freed buffers, which results in a BUG() from a double kfree().
However we still have an issue of possibly freeing buffers before the
hardware is done with them, so use ref counting to address that as well.

Change-Id: Icb72eaf81f09581f4be502dadb94f6bb3cb2cd8f
Signed-off-by: Jeffrey Hugo <jhugo@codeaurora.org>