qaic: Add power telemetry metrics
Its possible for us to query the device for a number of metrics.  We start
with adding a simple implementation which initializes a hwmon device at
the point where we can start issuing queries for these metrics.  We expose
a small set of power metrics to start.  For simplicity, we provide a
dummy implementation which stubs out the reads/writes of the sysfs files.
This allows userspace to implement their handling of these metrics with
dummy data, while the plumbing for talking to the device to get the real
data comes later.

The following metrics are exposed at this time:

power1_input - returns the instantanious power consumption of the device
at the board level.

power_level - returns the current configured power level.  Allows the level
to be configured to a different level.  The power levels are: (1) full power,
(2) reduced power, and (3) minimal power.  Configuring a different power
level impacts the permitted power consumption of the device, and thus may
impact the performance of the device.

throttle_percent - returns the percent of performance throttling the device
is imposing to meet the configured limits - ie thermal, power, etc limits.
Value is 0 to 100 (%).

Change-Id: Iabf2a0c0966fd1d76742962e2eab170c05f678b2
Signed-off-by: Jeffrey Hugo <jhugo@codeaurora.org>