qaic: Fix uninitialized variable in commit_execute()
The two_copy flag could be used in an uninitialized state.  Fix this by
initialized with a default value consistent with the intent of the flag.
That then exposes the issue that our two_copy check logic was fundamentally
flawed, and we were using garbage on the stack to always hit the two_copy
case.  Therefore, fix the logic as well.  With both fixes, we can actually
see the optimized one copy case being used, except when two copy is
necessary.

Change-Id: Id1ffb0ab89cf5c3e9bbce6fbe483ca63b26808be
Signed-off-by: Jeffrey Hugo <jhugo@codeaurora.org>