qaic: Add thermal telemetry metrics
Add the thermel telemetry metrics that we expect to be able to query, with
dummy implementations.

From the board (temp1), we can get the current temp, and historical highest
from the current device boot.

From the SoC (temp2), we can get the current temp, the historical highest,
the soft thermal limit where the hardware will start throttling, and also
the hard thermal limit where the device will shutdown.  We can also set the
soft and hard thermal limits.

From the DDR (temp3), we can tell if it is currently overheated.

Change-Id: I20e35c20165d2b0dc82d7f8fda275f508fc4ac97
Signed-off-by: Jeffrey Hugo <jhugo@codeaurora.org>