qaic: Ignore pending signals when sending cleanup messages to device
If the user was killed, the process will have a pending signal while it is
cleaned up.  Thus when the open file descriptors are closed, we won't wait
for the device's response to our cleanup request (the terminate command).
This can kick off a race condition, where the device is cleaning up and
stopping the use of resources we've given it, while we start cleaning up
those same resources.  Its possible we will release the resources before
the device is done using them, which could result in memory corruption.

To handle this, add the ability to send a message to the device while
ignoring pending signals, and use this ability only when doing cleanup so
that we properly give the device an oppertunity to stop using the
resources before we start our cleanup.

Change-Id: I12affd6a639a8091012d61c96000593dd9663a8b
Signed-off-by: Jeffrey Hugo <jhugo@codeaurora.org>