qaic: Support fatal error handling
When fatal error happens, system reboots and we land into PBL. PBL sends
a system error. MHI need to be reset in such situations.

While doing that, also move qaic_dev_reset_clean_local_state() a bit up,
so that we do not need to have a prototype for it before it's usage in
reset_mhi_work_func().

Change-Id: Ic06d915760892b81dcede7962963e4154d8794cf
Signed-off-by: Jeffrey Hugo <jhugo@codeaurora.org>