qaic: cleanup dbc buffer save
We create dbc buffer during encode activation step. If activation
command is successful then we need to keep this buffer saved. Currently,
we save this buffer in qaic_manage(), based on resource.status. resource
is a local variable and it's status is populated only during activation
command, and hence it could be stale for some other manage command.
Therefore, best place to save these buffer is decode_activate().

Change-Id: I9c5a49bd50691d266716c2e1bc7c47af0c56b999
Signed-off-by: Pratyush Anand <pratanan@codeaurora.org>