qaic: Handle MHI SYSERR
The device will trigger the MHI SYSERR state if it has encountered a
serious error and requires the MHI link to be reset to recover.  The
device will not preserve any state when this occurs, so we need to notify
clients and cleanup any resources.  When the MHI link is reestablished, we
reinit for normal operations.

Change-Id: I2e5b62772a019ed15fe42a275392a273d4e88ccf
Signed-off-by: Jeffrey Hugo <jhugo@codeaurora.org>