qaic: Make the user API a unique scope
In some kernel configurations, there is already a "struct sem", therefore
the qaic driver causes a symbol conflict.  The API to qaic should be in a
unique "scope" so that it does not conflict with anything else.  Update
API so that everything has a presix of "qaic" which is intended to be
unique.

Change-Id: I5b570ea95b57745cbbbc522dd577c5269f604285
Signed-off-by: Jeffrey Hugo <jhugo@codeaurora.org>