qaic: Implement trace support
Implement traces for the manage ioctl errors so that the exact root cause
can be identified when there are multiple paths which return the same error
code.  We use trace because some of these calls can be performance critical
and it will be useful to correlate these events with the data path in the
future.

Change-Id: I7bdda0fd3025d7a5e2499d9dc9b21ea3d96005f9
Signed-off-by: Jeffrey Hugo <jhugo@codeaurora.org>