qaic: Implement mhi link status callback
When the MHI core thinks that it detects that the link has gone down, it
will call back to the controller for a confirmation.  This is done via the
link_status() callback.  Our dummy implementation always returns
"all good", which causes the MHI core to proceed, which may result in bad
data.  Instead, implement a check that will return if the link is up or
down.

Change-Id: Iaf2cd1c7e799090eb29eff3bc56253249c145fce
Signed-off-by: Jeffrey Hugo <jhugo@codeaurora.org>