qaic: Batch submission of execute and mem ioctl
We will always be submitting all memory request and also execute
request IOCTL for one inference back to back. So, better to do that
using one IOCTL.
It will help user space to allocate memory for one inference in one go,
and also to submit execution for all of them together. It saves
overhead of multiple system call and also helps user space for minimal
book keeping in case of failure of one memory or exec request from batch.

Change-Id: I0e1ea241b0aa13e89ecb34d5e5a9c3f45551810f
Signed-off-by: Pratyush Anand <pratanan@codeaurora.org>