qaic: Add support of variable-length NNC messages
Currently all NNC messages are handled at maximum message size of 16KB,
which is too large for most of the use cases while too small for DMA
transactions when IOMMU is not enabled. With this patch the receving
end reads the message header first and then read again for the rest of
the message according to the length field in message header. This patch
sets a 4KB cap for most cases execept for messages with DMA transactions
which could be as large as 64KB.

Change-Id: I2f6986db6719ec91d0e2e63a6e5d2509ad291463
Signed-off-by: Fan Wu <wufan@codeaurora.org>