mhi_bus: dev: uci: Fix read() race condition
If two processes are calling read() on the same MHI device at the same
time, they might hit a race condition because the channel lock is unlocked
for a time.  This could allow both processes to cache a pointer to the
same buffer, which could later cause memory corruption if both processes
requeue the same buffer.

We have to unlock the lock because its a spinlock and copy_to_user() can
sleep, so create a mutex to serialize read accesses to the same channel.

Change-Id: I96426abfaf164e5b085690b08532f6177f7b002b
Signed-off-by: Jeffrey Hugo <jhugo@codeaurora.org>