qaic: Make sure to initialize xfer_done of mem handle
The xfer_done completion object is not being initialized until the mem
handle is queued to the hardware.  If the handle is never queued to the
hardware, but a wait_exec ioctl is issued on it, the uninitialized object
can cause a kernel crash.  If we make sure to initialized the object when
the handle is allocated, it can never be used uninitialized.

Change-Id: I4fc9307469a8a1a73195d06c127627acd693aa20
Signed-off-by: Jeffrey Hugo <jhugo@codeaurora.org>