qaic: Fix event ring config
The aic100 event ring is not dedicated to a specific channel, but it is
incorrectly configured as such.  This causes the MHI stack to reset the
event ring when a channel is being unprepared, which can cause the ring
to be corrupted, and trigger invalid duplicate events if the unprepare
processing completes before the event ring tasklet completes.  By
correctly configuring the event ring as non-dedicated, we avoid the race
and no longer see memory corruption and crashes under long running stress
tests.

Change-Id: I3a993d12a444f39a3b7a55e15fb7d698bfec6278
Signed-off-by: Jeffrey Hugo <jhugo@codeaurora.org>