qaic: fix possible list corruption race
A list corruption report has occured from the list_del in free_usr() from
the unlocked path in device_release().  Attempts to repro the issue have
failed, however a code inspection suggests that the assumption that
free_usr() could be called in an unlocked context in specific scenarios
may be invalid.  If multiple users for the same device are able to invoke
free_usr() concurrently through the unlocked path, they may race and cause
list corruption since list_del() is not atomic.

Fix the issue based on the code inspection by using the lock anytime
the list state is changed.

Change-Id: I8aa000630723b7e78201a9915dc9cd38ba53fdeb
Signed-off-by: Jeffrey Hugo <jhugo@codeaurora.org>