qaic: Reset lost buf flags on init
If the device has crashed, we are not resetting the lost buf flags for
control and telemetry communication.  If the flag was set before the
device crashed, we will attempt to consume the lost buf after the device
has reset, which is not correct.  Therefore on control/telemetry init,
make sure the flag is set to the correct state.

Change-Id: I365853e54b5b5f7fc16f9bccbea9d1a8436d988b
Signed-off-by: Jeffrey Hugo <jhugo@codeaurora.org>