qaic: Allow the reset process to exit the reset state upon request
If we take a non-fatal error from MHI, we rely on the MHI state to go back
to AMSS and cause the control channel to probe as the signal to exit the
reset state.  However, if the device signals a fatal error before this,
we don't exit the reset state, and will ignore the fatal error since we
assume that a reset is already being processed.  This will leave the
device in an uninitialized state.

We can fix this by allowing the client initiating the reset to indicate
that the reset state should be cleared once the reset processing is done.

Change-Id: Id36c090a13944c3c8f2803188dae46804187054e
Signed-off-by: Jeffrey Hugo <jhugo@codeaurora.org>