qaic: Change qaic device node name
It turns out that docker treats the ":" as part of a display, so its not
possible to pass a qaic device into a docker container.  Thus, we need to
change the device node name to support docker.

While we could choose a different delimiter, that would differentiate from
the "standard" PCI address format used elsewhere, which makes the bus
address in the device name less self-apparent.

Instead, lets follow the example of the DRM subsystem, and the Habana Labs
driver by using the device minor number as the unique identifier.

Example:
from - qaic_aic100_0000:1a:00.0
to - qaic_aic100_1

To make a link back to the PCI device which this qaic device is related to,
follow the DRM example of setting the device parent.

Finally, device_create() can set up the drvdata for us, allowing us to
remove the need to manually set that.

Change-Id: If97a1fc48534c1cd6e8e0289d9d59c732aba8359
Signed-off-by: Jeffrey Hugo <jhugo@codeaurora.org>