qaic: Do page shift before extracting handle and dbc_id
We store both memory handle and dbc_id on same variable handle. First
we mask the memory handle on lower 32-bit and dbc_id on upper 32-bit
and secondly we left shift the entire variable by PAGE_SHIFT. In the
current code we first mask out both memory handle and dbc_id and
then perform right shift on both individually. With the current
approach if the value of memory handle is stored in more than 16-bit
and PAGE_SHIFT is 16 the code will break.

Change-Id: Ia00d23891340a907a614285726a6076c8aaf74bd
Signed-off-by: Pranjal Ramajor Asha Kanojiya <pkanojiy@codeaurora.org>