qaic: Reserve pages from last iteration
In the last iteration of page allocation, we do a tail fixup, and then
exit before reserving the pages from the lasst allocation.  This means
that the pages could possibly be swapped out while the DMA engine is
attempting to access them, resulting in corrupt data sent to the device.

Change-Id: I238c6bc02431d544cd4105021ae711de594fb99b
Signed-off-by: Pranjal Ramajor Asha Kanojiya <pkanojiy@codeaurora.org>