accel/qaic: Add debugfs framework
Create a debugfs framework to expose device information.

Add debugfs nodes to expose each DMA brdge channels(DBC) fifo size which is
the maximum number of DMA requests that can be queued on that DBC and
queued is the number of DMA requests in the fifo.
Example of a dbc debugfs:
	/sys/kernel/debug/accel/0/dbc000/fifo_size
	/sys/kernel/debug/accel/0/dbc000/queued

Register for QAIC_LOGGING mhi channel. This is used by device to send boot
logs. Make this boot log available via debugfs.
Example of boot log debugfs:
	/sys/kernel/debug/accel/0/bootlog

Change-Id: I0040bfa9454a0efd3816c1dd1788299ec7399e90
Signed-off-by: Pranjal Ramajor Asha Kanojiya <quic_pkanojiy@quicinc.com>