accel/qaic: Prevent NULL dereference in tracepoint
During get_cntl_version(), qaic_manage() is called with a NULL usr.
This is handled in the normal control flow, however, if an error is
encountered the tracepoints will try to log the user handle causing a
NULL dereference.

Check the usr for NULL before trying to find its handle.

Change-Id: Ie076164f9e609c67e93c25129ac7b970240e0472
Signed-off-by: Carl Vanderlip <quic_carlv@quicinc.com>