net: qrtr: ns: Add force_remove module parameter
ns is part of the qrtr.ko, and also opens up a kernel socket to qrtr.
This causes the qrtr.ko to automatically have a +2 reference count, and
blocks any attempt to remove the module. Add a module parameter that when
written to, will cause the ns socket to be closed, thus allowing the
module to be removed. Significant changes to the code are needed for a
clean rmmod method.

Change-Id: I1c3ef21007a27297547bec73f5f55de43645a51b
Signed-off-by: Jeffrey Hugo <quic_jhugo@quicinc.com>