accel/qaic: Update SSR handling
The subsystem restart (SSR) handling is changing in firmware.

Currently QSM will essentially deactivate the workload, and reactivate
it. This makes the SSR event mostly transparent from the user
perspective.

However, this doesn't work when the error that triggers SSR does not
allow for a reactivate, or the network retains some kind of state from
input to input. For the former, SSR cannot complete. For the latter, the
issue can be silent but the end user will not get the correct output.

Instead, SSR behavior in the firmware will involve a deactivate, but not
a reactivate. The end user will need to be aware of SSR, and decide the
course of action.

For the KMD, this means we can release the DBC at the start of SSR and
no longer need to hold it in "suspend" for the user.

Change-Id: I9d84d8a05c82af9ee2711f64cc929f0276313849
Signed-off-by: Jeffrey Hugo <quic_jhugo@quicinc.com>