backport: Fix class_create call in accel_helpers
Backport patch to deal with class_create was changed to match on calls
with only one argument. It was assumed that since the new class_create
occurred post drm_accel, that accel_helpers wouldn't need to swap to
the new version. However, if DRM_ACCEL isn't configured, the
backported accel framework still needs to be used.

In addition to changing the call to the single argument, the SmPL
patch currently will only match against identifiers, and not the
string as used currently. Create a definition that maps to the "accel"
name so that the SmPL patch will apply here too.

Change-Id: I35e60d9ebd1f2b87a53374f9a1e3c40fd4efc14b
Signed-off-by: Carl Vanderlip <quic_carlv@quicinc.com>