accel/qaic: Set number of DBCs allocated per device
The driver currently supports 4 devices, and they can differ in
configurations, including the number of DBCs. Currently the DBC count
is set to 16 regardless of the device.

Add definitions for the number of DBCs per device. Use them as private
data to set the number of DBCs to be allocated for each device.

Change-Id: I89a470a38779cbc9e0d2b9db28851917a1d13122
Signed-off-by: Youssef Samir <quic_yabdulra@quicinc.com>