accel/qaic: Update PCI Device ID definitions
The PCI_DEVICE_DATA(vend, dev, data) macro expects the dev field to be
of a specific format. The reason for choosing this macro is that it
sets the driver_data field of the pci_device_id struct.

Change the Device ID definitions from PCI_DEV_AICxxx
to PCI_DEVICE_ID_QCOM_AICxxx.

Change-Id: I3da1ff144af6674bfa4f1fa6c5a0ad90712291d7
Signed-off-by: Youssef Samir <quic_yabdulra@quicinc.com>