accel/qaic: Add config structs for supported cards
As the number of cards supported by the driver grows, their
configurations will differ. The driver needs to become more dynamic
to support these configurations. Currently, each card may differ in
the exposed BARs, the regions they map to, and the DBC count.

Create config structs for each card, and let the driver configure the
qaic_device struct based on the configurations passed to the driver.

Change-Id: I0a7a60f3b35bf549546240f369c8498cefef9f41
Signed-off-by: Youssef Samir <quic_yabdulra@quicinc.com>