accel/qaic: Update SSR dump header
SSR dumps will be going through the same devcoredump framework as Sahara
dumps. This means that userspace will need to be able to parse both.
Update the SSR dump header to match the structure of the Sahara dump
header so that userspace can parse both, and identify any future
changes.

Change-Id: I88070c6c7b837e99866da24846e98b958006b95e
Signed-off-by: Jeffrey Hugo <quic_jhugo@quicinc.com>